homework_36/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
passlib
aiofiles
pydantic
python-jose